--- app/src/main/java/app/krunal3kapadiya/facts/Injection.kt
@@ -1,11 +1,0 @@
-package app.krunal3kapadiya.facts
-
-object Injection {
-    /**
-     * fact view model
-     */
-    @JvmStatic
-    fun provideFactsViewModel(): FactsViewModelFactory {
-        return FactsViewModelFactory()
-    }
-}
