+++ app/src/main/java/app/krunal3kapadiya/facts/network/models/Facts.kt
@@ -8,7 +8,10 @@ import com.google.gson.annotations.SerializedName
  * @date 14,April,2019
  */
 
-data class Facts(
-    @SerializedName("title") val title: String,
-    @SerializedName("rows") val rowsList: List<FactsRows>
+data class Facts(@SerializedName("title") val title: String, @SerializedName("rows") val rowsList: List<FactsRows>)
+
+data class FactsRows(
+    @SerializedName("title") var title: String?, @SerializedName("description") var description: String?, @SerializedName(
+        "imageHref"
+    ) var imageHref: String?
 )
