+++ app/src/main/java/app/krunal3kapadiya/facts/utils/Utility.kt
@@ -1,0 +1,40 @@
+package app.krunal3kapadiya.facts.utils
+
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import android.os.Build
+
+class Utility {
+    companion object{
+        // checking internet
+        @Suppress("DEPRECATION")
+        fun isInternetAvailable(context: Context): Boolean {
+            var result = false
+            val cm = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager?
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                cm?.run {
+                    cm.getNetworkCapabilities(cm.activeNetwork)?.run {
+                        result = when {
+                            hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> true
+                            hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> true
+                            hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET) -> true
+                            else -> false
+                        }
+                    }
+                }
+            } else {
+                cm?.run {
+                    cm.activeNetworkInfo?.run {
+                        if (type == ConnectivityManager.TYPE_WIFI) {
+                            result = true
+                        } else if (type == ConnectivityManager.TYPE_MOBILE) {
+                            result = true
+                        }
+                    }
+                }
+            }
+            return result
+        }
+    }
+}
