+++ app/src/main/res/layout/activity_facts_list.xml
@@ -5,12 +5,11 @@
     android:id="@+id/vf_list"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.factsList.FactsListActivity">
+    tools:context=".ui.factslist.FactsListActivity">
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
         android:id="@+id/swipe_refresh"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        style="@style/MatchContentStyle"
         android:backgroundTint="@color/colorAccent"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
@@ -19,17 +18,14 @@
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_fact"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
+            style="@style/MatchContentStyle" />
     </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
+    <androidx.constraintlayout.widget.ConstraintLayout style="@style/MatchContentStyle">
 
         <ProgressBar
             android:id="@+id/pb_loading"
-            android:layout_width="0dp"
+            android:layout_width="@dimen/size_zero"
             android:layout_height="wrap_content"
             app:layout_constraintBottom_toTopOf="@+id/txt_loading"
             app:layout_constraintEnd_toEndOf="parent"
@@ -39,7 +35,7 @@
 
         <TextView
             android:id="@+id/txt_loading"
-            android:layout_width="0dp"
+            android:layout_width="@dimen/size_zero"
             android:layout_height="wrap_content"
             android:gravity="center"
             android:text="@string/loading"
@@ -50,9 +46,7 @@
             app:layout_constraintTop_toBottomOf="@+id/pb_loading" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
+    <androidx.constraintlayout.widget.ConstraintLayout style="@style/MatchContentStyle">
 
         <androidx.appcompat.widget.AppCompatImageView
             android:id="@+id/img_no_internet"
@@ -80,7 +74,7 @@
 
         <androidx.appcompat.widget.AppCompatTextView
             android:id="@+id/txt_no_internet"
-            android:layout_width="0dp"
+            android:layout_width="@dimen/size_zero"
             android:layout_height="wrap_content"
             android:layout_marginTop="@dimen/vertical_internet_txt_margin"
             android:layout_marginBottom="@dimen/vertical_internet_txt_margin"
