+++ app/src/main/res/layout/row_lists.xml
@@ -2,16 +2,14 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
+    style="@style/MatchWrapContentStyle"
     android:orientation="vertical"
-    android:padding="16dp"
-    tools:context=".ui.factsList.FactsListAdapter">
+    android:padding="@dimen/list_padding"
+    tools:context=".ui.factslist.FactsListAdapter">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/txtTitle"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        style="@style/MatchWrapContentStyle"
         android:layout_marginEnd="@dimen/text_margin"
         android:maxLines="1"
         android:textAppearance="@style/Base.TextAppearance.AppCompat.Title"
@@ -21,8 +19,7 @@
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/txtDescription"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        style="@style/MatchWrapContentStyle"
         android:layout_marginEnd="@dimen/text_margin"
         android:maxLines="2"
         app:layout_constraintEnd_toEndOf="parent"
