+++ app/src/main/res/values/dimens.xml
@@ -5,4 +5,6 @@
     <dimen name="img_wifi_size">108dp</dimen>
     <dimen name="button_wifi_width">96dp</dimen>
     <dimen name="vertical_internet_txt_margin">16dp</dimen>
+    <dimen name="size_zero">0dp</dimen>
+    <dimen name="list_padding">16dp</dimen>
 </resources>
